Add in-Livebook VM runtime
This adds a runtime so that it's possible to run code from within the
same VM that runs the Livebook application.

This is useful for Nerves where there's only one Erlang VM running and
it runs the interesting libraries and applications that are useful for
Nerves Livebooks. While it's possible to run child VMs and compile pure
Elixir libraries on device, non-Elixir code can't be compiled by design
(no C compiler on device), so the very interesting feature of being able
to pull hex dependencies down in Livebooks is not feasible barring
binary packages.

This commit is a WIP since it's a departure from the much more elegant
handling of Livebook code and I'm looking for feedback especially if
there's a simpler way. In retrospect, the name "local" for this runtime
should probably be changed. There's also a hack to make this method the
default, since it's awkward for it to not be the default on a Nerves
device.