Fix tests
When we send a message that a node is down to the NodePool, that
message is put into the NodePool's mailbox. The NodePool then
checks the mailbox for messages and acts on them.

However, we do not have any guarantee on *which* message is selected
by the GenServer which is acted on. Thus, if you send a call that
depends on the contents of a previously sent message, there's is a
good chance, that the call is acted upon first, and you don't get an
expected result.

Hence, in our tests, we are waiting one microsecond before we send our
call, so that our `:nodedown` message is fully processed (The GenServer
sends itself another message which needs to be processed before we call
`get_name`).