Add :notebook clause for opts_to_config/2
@ryg-git 

I've added the :notebook clause for opts_to_config/2 (line 217).
Since we want to open the notebook after changing the root path, and the `--open` flag may not be entered by user when using `--notebook`, I'm adding `{:open, true}` in the `opts` list. I've removed the if-else clause for `opts[:notebook]` added earlier in open_from_options/1, since that would be incorrect as the `:notebook` flag is not of boolean type.