Fix macOS menu bar
Previously we were removing the "Hide [app]", "Hide Others" and "Show all" as they weren't working out of the box. Turns out all we needed to do is:

    - :wxFrame.connect(frame, :command_menu_selected)
    + :wxFrame.connect(frame, :command_menu_selected, skip: true)

as this will make it so the original (system-level) event handler is
fired. Then all we need to do is ignore these events in the Elixir land.

Given by default the app name is "Erlang", we need to update the menu
item labels accordingly. (For "Hide [app]" and "Quit [app]".)