Fix launching Windows desktop app
On Windows we have erl.exe in two directories: OTP/bin/erl.exe and
OTP/erts-VSN/bin/erl.exe. The executable is identical but what is
different is the latter, the erts one, contains erlexec.dll which needs
to be present for the VM to start.

Interestingly, regardless of OS, the erl[.exe] executable always
prepends the OTP/erts-VSN/bin and OTP/bin directories to the PATH:

    iex> System.get_env("PATH") |> String.split(":") |> Enum.take(2)
    ["/Users/wojtek/.asdf/installs/erlang/25.1.2/erts-13.1.2/bin",
     "/Users/wojtek/.asdf/installs/erlang/25.1.2/bin"]

Manually prepending the erts bin directory to the PATH, which makes the
erl.exe that is to be launched to be the erts one, fixed the issue.