Rebuild Mac app
Before this patch, our app consisted of basically two processes, the
launcher, and the BEAM. Importantly, the GUI was running in the latter
one, which seem to go against the grain of building Mac apps. For
example, the native function to report the .app bundle path
(`Bundle.main.bundlePath` in Swift) returns the wrong one which then
caused many bugs in wxErlang and wxWidgets.

In this patch we're introducing a new way of bundling apps. We're
building OTP ourselves (including its dependencies, at the moment just
openssl) and statically linking it to the launcher program which then
starts the VM.

Speaking of building OTP ourselves, in this patch we're cross-compiling
it on both architectures. Since we can now create universal build
locally (e.g. in public GitHub runners) we don't need a separate ARM Mac
in the cloud anymore.

Before this patch we wanted to avoid any "runtime" in the bundler code
Now we're taking the opposite approach, we're exposing functions that
help building the app on both the native and Elixir side and communicate
back and forth. (The communication now happens inside the same OS
process and thus is much more efficient.) To mark the occasion we're
renaming AppBundler to ElixirKit. Here are the APIs:

On the Swift side:

```
ElixirKit.start()
ElixirKit.startAsync()
ElixirKit.publish(name, data)
ElixirKit.publisher(name)
ElixirKit.subscribe(name, queue:, block:)
```

and on Elixir side:

```
ElixirKit.publish(name, data)
ElixirKit.subscribe(name)
```

See `elixirkit/demo` for a sample project.

Previously we tried abstracting away all the app configuration, for
example we were automatically generating an `Info.plist`. That approach
worked pretty well but was obviously limiting. Now, we're asking users
to bring their own `Info.plist.eex` (with access to `@release`) which
gives them full flexibility. We're also adding a `:resources`
configuration option to easily copy assets to the app bundle.

On the flip side, previously when the VM crashed we were able to catch
that and show an error dialog. Now we can't because when the VM crashes,
that's it. In the future, we could introduce a separate supervisor
program that monitors the VM and acts accordingly. At the moment,
we changed `Livebook.Config.abort!` to send a message to the frontend to
display an error dialog and then quit.

Similarly, we now cannot simply redirect stdout/stderr to a log file. I
think we should configure a logger file backend and change `IO.puts` to
`Logger.*` calls.

At the moment we're not cross-compiling wxWidgets and thus our OTP
doesn't not have the wx app and it's dependencies: debugger, observer,
cdv, et. We don't include jinterface and odbc.

This PR does not address Windows side of things yet.