Widen the inputs for the auth screen
Expand the input fields to the full width of the box.

The use of `items-center` in the form squishes both the button and text field to the middle of the bounding area.
Removing the form class to let the button expand to full width looks a little awkward.
This seems like a decent enough compromise.