the naming of data Transform Cell + some grammar
IMO Data transform cell should be one of: data Transform Cell or Data Transform Cell or even data transform cell. I say this since if you capitalise Data, it assumes that Data has been declared now as something special, and so needs a name (or apriori, in this document or elsewhere). But it is the Transform Cell that is infact special in this context. Later in the document you have Smart Cells. This makes sense to me as the set of cells which have smart properties. And you have Chart cell - which I can judge to be the short-style name of Chart-configuration cell. i.e. it is an auxiliary activity upon the main stuff, the data. But, in the context of this document, Data is confusing.  Anyhow, I chose one of these i.e. The smart data Transform Cell

In the near future = shortly;    TL;DR = in short