Mobile: Prevent zoom on input focus in WebKit mobile browsers
There is unfortunately not a great way to do this. Apparently, you can
increase the textarea font-size to 16px, but I don't think we want to do
that. Adding `maximum-scale=1.0` to the viewport meta tag works well in
WebKit browsers, but some mobile browsers, like Chrome on Android, will
prevent the user from zooming at all if that is present. Instead of
doing that, this hook does its best to detect that we're in a WebKit
browser and only adds the meta attribute if so.

Addresses part of #2011.