Rename "running sessions" to "running notebooks"
This is purely a UX change to align the list of running "notebooks"
(sessions) with the button to create a "New notebook".

Notebooks and Sessions are very much distinct concepts in the backend
and it doesn't make sense to combine them. From the module docs a
notebook is a user editable document that maps to a plain file and a
notebook session is a server process that holds a specific notebook
instance.

Resolves #2872

Discussion: https://github.com/livebook-dev/livebook/discussions/2871