Use a non-hidden file to track if escript assets are extracted
When tmpdir is cleaned on macOS, files are removed, directories are
kept, so we use an extra file as a flag to easily check if the files
are extracted. However, files starting with dot are not removed as
part of the cleanup, so we need to use a regular name.